+++ ariesk/utils/bloom_filter.pxd
@@ -16,7 +16,7 @@ cdef class BloomFilter:
 
     cdef add(self, npc.uint8_t[:] seq)
     cdef bint contains(self, npc.uint8_t[:] seq)
-    cdef bint contains_hvals(self, npc.uint64_t[:] hvals)
+    cdef bint contains_hvals(self, npc.uint32_t[:] hvals)
     cpdef int union(self, BloomFilter other)
     cpdef int intersection(self, BloomFilter other)
 
@@ -29,10 +29,10 @@ cdef class BloomGrid:
     cdef public npc.uint64_t[:, :] col_hashes, row_hashes
 
     cdef add(self, npc.uint8_t[:] seq)
-    cdef npc.uint64_t[:] _get_hashes(self, npc.uint8_t[:] seq)
+    cdef npc.uint32_t[:] _get_hashes(self, npc.uint8_t[:] seq)
     cdef bint array_contains(self, npc.uint8_t[:] seq)
-    cdef bint array_contains_hvals(self, npc.uint64_t[:] hvals)
+    cdef bint array_contains_hvals(self, npc.uint32_t[:] hvals)
     cdef npc.uint8_t[:] grid_contains(self, npc.uint8_t[:] seq)
-    cdef npc.uint8_t[:] grid_contains_hvals(self, npc.uint64_t[:] hvals)
+    cdef npc.uint8_t[:] grid_contains_hvals(self, npc.uint32_t[:] hvals)
     cdef npc.uint8_t[:] count_grid_contains(self, npc.uint8_t[:] seq)
-    cdef  npc.uint8_t[:] count_grid_contains_hvals(self, npc.uint64_t[:, :] hvals)
+    cdef  npc.uint8_t[:] count_grid_contains_hvals(self, npc.uint32_t[:, :] hvals)
